split db func, fix attendance display
