+++ test_data.txt
@@ -1,0 +1,147 @@
+16:47:23.663179
+Belle: 0.3587531875295964
+
+16:47:24.111179
+Belle: 0.35502348621510044
+
+16:47:24.523212
+Belle: 0.3455767407345685
+
+16:47:24.930206
+Belle: 0.3495531327031432
+
+16:47:25.328210
+Belle: 0.377884137029053
+
+16:47:25.727210
+Belle: 0.3646718197738056
+
+16:47:26.134210
+Belle: 0.3413029146995752
+
+16:47:26.527890
+Belle: 0.34675202765438096
+
+16:47:26.923891
+Belle: 0.3473310874951758
+
+16:47:27.351892
+Belle: 0.3601168698765336
+
+16:47:27.775922
+Belle: 0.35521684914108076
+
+00:34:41.426910
+Belle: 0.467921433708988
+
+00:34:41.835910
+Belle: 0.46479164152506336
+
+00:34:42.227912
+Belle: 0.45732229176083955
+
+00:34:42.619878
+Belle: 0.4502332598083385
+
+00:34:43.023909
+Belle: 0.4759814108127348
+
+00:34:43.413912
+Belle: 0.47736630894717064
+
+00:34:43.804268
+Unknown: 0.507421129927732
+
+00:34:44.197455
+Belle: 0.4817152152580633
+
+00:34:44.590129
+Belle: 0.43083732432688077
+
+00:34:44.980162
+Belle: 0.384916488591851
+
+00:34:45.373141
+Belle: 0.2962326202696426
+
+00:34:45.780130
+Belle: 0.29144037857218197
+
+00:34:46.179130
+Belle: 0.2935769081084843
+
+00:34:46.603121
+Belle: 0.3103726255853578
+
+00:34:46.995331
+Belle: 0.3119750689615828
+
+00:34:47.394299
+Belle: 0.32011775466700026
+
+00:34:47.785299
+Belle: 0.3496760276030264
+
+00:34:48.191792
+Belle: 0.33378289895867314
+
+00:34:48.597767
+Belle: 0.3099477787090603
+
+00:34:48.991796
+Belle: 0.3060433331254786
+
+00:34:49.460793
+Belle: 0.33189845268623785
+
+00:34:49.855760
+Belle: 0.32214490231060083
+
+00:34:50.248793
+Belle: 0.305429172967294
+
+00:34:50.638761
+Belle: 0.32164098256448814
+
+00:34:51.030152
+Belle: 0.3150345692981525
+
+00:34:51.421153
+Belle: 0.33803708025474827
+
+00:34:51.818121
+Belle: 0.3315444791783201
+
+00:34:52.210636
+Belle: 0.3287463470074036
+
+00:34:52.628949
+Belle: 0.302607101945604
+
+00:34:53.035979
+Belle: 0.33530357412998807
+
+00:34:53.434948
+Belle: 0.339887683636815
+
+00:34:53.835978
+Belle: 0.39583738081463954
+
+00:34:54.244431
+Belle: 0.4342329244406703
+
+00:34:55.433258
+Belle: 0.49899532106423966
+
+00:34:55.824227
+Belle: 0.356374767463089
+
+00:34:56.234028
+Belle: 0.3528575875642102
+
+00:34:56.652028
+Belle: 0.3163148435128022
+
+00:34:57.106030
+Belle: 0.34269785107866935
+
