+++ test_distances.txt
@@ -4558,3 +4558,161 @@ Names: ['Belle', 'Human1', 'Human2', 'Human3', 'Human4', 'Human5', 'Human6', 'Hu
 Distances: [0.74315839 0.8484495  0.76229586 0.65281583 0.70232611 0.81175938
  0.74270475 0.76556402 0.72970547 0.45158231]
 
+Belle Detected: 16:47:23.663179
+Names: ['Belle', 'Human1', 'Human2', 'Human3', 'Human4', 'Human5', 'Human6', 'Human7', 'Human8', 'Ike']
+Distances: [0.35875319 0.88679462 0.90720513 1.00128375 0.86465716]
+
+Belle Detected: 16:47:24.111179
+Names: ['Belle', 'Human1', 'Human2', 'Human3', 'Human4', 'Human5', 'Human6', 'Human7', 'Human8', 'Ike']
+Distances: [0.35502349 0.88008801 0.9052047  1.00057127 0.84347001]
+
+Belle Detected: 16:47:24.523212
+Names: ['Belle', 'Human1', 'Human2', 'Human3', 'Human4', 'Human5', 'Human6', 'Human7', 'Human8', 'Ike']
+Distances: [0.34557674 0.87596296 0.91117157 0.99732571 0.84900611]
+
+Belle Detected: 16:47:24.930206
+Names: ['Belle', 'Human1', 'Human2', 'Human3', 'Human4', 'Human5', 'Human6', 'Human7', 'Human8', 'Ike']
+Distances: [0.34955313 0.87620442 0.90744639 0.99765416 0.84418803]
+
+Belle Detected: 16:47:25.328210
+Names: ['Belle', 'Human1', 'Human2', 'Human3', 'Human4', 'Human5', 'Human6', 'Human7', 'Human8', 'Ike']
+Distances: [0.37788414 0.86438975 0.89961792 1.00158505 0.85901114]
+
+Belle Detected: 16:47:25.727210
+Names: ['Belle', 'Human1', 'Human2', 'Human3', 'Human4', 'Human5', 'Human6', 'Human7', 'Human8', 'Ike']
+Distances: [0.36467182 0.86525267 0.90292903 1.00646748 0.85408399]
+
+Belle Detected: 16:47:26.134210
+Names: ['Belle', 'Human1', 'Human2', 'Human3', 'Human4', 'Human5', 'Human6', 'Human7', 'Human8', 'Ike']
+Distances: [0.34130291 0.85895861 0.89968262 0.99985679 0.84640922]
+
+Belle Detected: 16:47:26.527890
+Names: ['Belle', 'Human1', 'Human2', 'Human3', 'Human4', 'Human5', 'Human6', 'Human7', 'Human8', 'Ike']
+Distances: [0.34675203 0.86975291 0.91714159 1.00473967 0.84941071]
+
+Belle Detected: 16:47:26.923891
+Names: ['Belle', 'Human1', 'Human2', 'Human3', 'Human4', 'Human5', 'Human6', 'Human7', 'Human8', 'Ike']
+Distances: [0.34733109 0.86205419 0.91290335 0.99567948 0.84309055]
+
+Belle Detected: 16:47:27.351892
+Names: ['Belle', 'Human1', 'Human2', 'Human3', 'Human4', 'Human5', 'Human6', 'Human7', 'Human8', 'Ike']
+Distances: [0.36011687 0.86381876 0.9148841  1.002578   0.85460221]
+
+Belle Detected: 16:47:27.775922
+Names: ['Belle', 'Human1', 'Human2', 'Human3', 'Human4', 'Human5', 'Human6', 'Human7', 'Human8', 'Ike']
+Distances: [0.35521685 0.86731712 0.91622543 1.00221832 0.84557961]
+
+Belle Detected: 00:34:41.426910
+Distances: [0.46792143 0.80523159 0.8435491  0.90233475 0.79343512]
+
+Belle Detected: 00:34:41.835910
+Distances: [0.46479164 0.80257073 0.83047306 0.9106839  0.80331228]
+
+Belle Detected: 00:34:42.227912
+Distances: [0.45732229 0.80315351 0.82480676 0.89734298 0.79616516]
+
+Belle Detected: 00:34:42.619878
+Distances: [0.45023326 0.79351994 0.8384154  0.90277213 0.77477706]
+
+Belle Detected: 00:34:43.023909
+Distances: [0.47598141 0.77375467 0.83300547 0.90575981 0.78782352]
+
+Belle Detected: 00:34:43.413912
+Distances: [0.47736631 0.77176261 0.8302864  0.85799684 0.77780617]
+
+Unknown Detected: 00:34:43.804268
+Distances: [0.50742113 0.76170783 0.82430145 0.87186797 0.79401643]
+
+Belle Detected: 00:34:44.197455
+Distances: [0.48171522 0.75021695 0.79441054 0.84120556 0.75399021]
+
+Belle Detected: 00:34:44.590129
+Distances: [0.43083732 0.7570211  0.81488847 0.86697132 0.74805601]
+
+Belle Detected: 00:34:44.980162
+Distances: [0.38491649 0.80400657 0.87504093 0.90032109 0.80107091]
+
+Belle Detected: 00:34:45.373141
+Distances: [0.29623262 0.84762754 0.88817402 0.95205624 0.86886761]
+
+Belle Detected: 00:34:45.780130
+Distances: [0.29144038 0.82730128 0.88292138 0.92980355 0.85604032]
+
+Belle Detected: 00:34:46.179130
+Distances: [0.29357691 0.83552828 0.88023929 0.9440298  0.85737525]
+
+Belle Detected: 00:34:46.603121
+Distances: [0.31037263 0.84626945 0.90102543 0.96243689 0.86020384]
+
+Belle Detected: 00:34:46.995331
+Distances: [0.31197507 0.83702847 0.90266762 0.95864933 0.86034208]
+
+Belle Detected: 00:34:47.394299
+Distances: [0.32011775 0.82713703 0.90021266 0.96169941 0.8625867 ]
+
+Belle Detected: 00:34:47.785299
+Distances: [0.34967603 0.82459166 0.89779976 0.94321022 0.85198105]
+
+Belle Detected: 00:34:48.191792
+Distances: [0.3337829  0.82007618 0.904137   0.93478276 0.84909145]
+
+Belle Detected: 00:34:48.597767
+Distances: [0.30994778 0.81637529 0.90036735 0.93560878 0.84748208]
+
+Belle Detected: 00:34:48.991796
+Distances: [0.30604333 0.82483524 0.89016041 0.9284815  0.85600527]
+
+Belle Detected: 00:34:49.460793
+Distances: [0.33189845 0.83064626 0.88811038 0.93791521 0.84894794]
+
+Belle Detected: 00:34:49.855760
+Distances: [0.3221449  0.82924006 0.88901246 0.93804441 0.8585235 ]
+
+Belle Detected: 00:34:50.248793
+Distances: [0.30542917 0.82080701 0.88195253 0.93825929 0.85609701]
+
+Belle Detected: 00:34:50.638761
+Distances: [0.32164098 0.83969591 0.91191299 0.96147742 0.85473984]
+
+Belle Detected: 00:34:51.030152
+Distances: [0.31503457 0.81901333 0.88733205 0.9351717  0.837097  ]
+
+Belle Detected: 00:34:51.421153
+Distances: [0.33803708 0.82786997 0.89508894 0.93322316 0.84365941]
+
+Belle Detected: 00:34:51.818121
+Distances: [0.33154448 0.82954058 0.88996679 0.93900155 0.84759319]
+
+Belle Detected: 00:34:52.210636
+Distances: [0.32874635 0.83055755 0.89998173 0.93463097 0.84331395]
+
+Belle Detected: 00:34:52.628949
+Distances: [0.3026071  0.82355123 0.88547814 0.94114713 0.84466237]
+
+Belle Detected: 00:34:53.035979
+Distances: [0.33530357 0.83116964 0.88325364 0.9299884  0.83954465]
+
+Belle Detected: 00:34:53.434948
+Distances: [0.33988768 0.82824699 0.87596162 0.93396932 0.84176259]
+
+Belle Detected: 00:34:53.835978
+Distances: [0.39583738 0.84641358 0.85835302 0.97955821 0.8677468 ]
+
+Belle Detected: 00:34:54.244431
+Distances: [0.43423292 0.79004538 0.84789752 0.94692317 0.84400026]
+
+Belle Detected: 00:34:55.433258
+Distances: [0.49899532 0.75997322 0.88674664 0.91143085 0.83023128]
+
+Belle Detected: 00:34:55.824227
+Distances: [0.35637477 0.78540207 0.8433192  0.94678754 0.82796048]
+
+Belle Detected: 00:34:56.234028
+Distances: [0.35285759 0.80554352 0.84924174 0.94978093 0.82320224]
+
+Belle Detected: 00:34:56.652028
+Distances: [0.31631484 0.84170233 0.89149093 0.97314804 0.85571165]
+
+Belle Detected: 00:34:57.106030
+Distances: [0.34269785 0.81733399 0.88251534 0.90872927 0.84325188]
+
